Secret.cs: Split file into Secret.cs and Secret`1.cs
... so that each class is represented by one single file.

Resolves: No entry
