+++ src/SecretSharingDotNetFx4.6.2.csproj
@@ -46,6 +46,7 @@
   <ItemGroup>
     <Compile Include="Cryptography\FinitePoint.cs" />
     <Compile Include="Cryptography\Secret.cs" />
+    <Compile Include="Cryptography\Secret`1.cs" />
     <Compile Include="Cryptography\ShamirsSecretSharing`1.cs" />
     <Compile Include="Cryptography\ShamirsSecretSharing`3.cs" />
     <Compile Include="Cryptography\SharesEnumerator.cs" />
